requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
flask-bcrypt==1.0.1
flask-limiter==4.1.1
psycopg2-binary==2.9.9
openpyxl==3.1.2
Pillow==10.3.0
pytesseract==0.3.10